compilation fix for 64 bit platforms (where size_t is not unsigned int)
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@44447 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ tests/strings/strings.cpp
@@ -613,7 +613,7 @@ void StringTestCase::WriteBuf()
     CPPUNIT_ASSERT_EQUAL(_T('f'), s[0u]);
     CPPUNIT_ASSERT_EQUAL(_T('o'), s[1]);
     CPPUNIT_ASSERT_EQUAL(_T('o'), s[2]);
-    CPPUNIT_ASSERT_EQUAL(3u, s.length());
+    CPPUNIT_ASSERT_EQUAL((size_t)3, s.length());
 
     {
         wxChar *p = s.GetWriteBuf(10);
@@ -624,7 +624,7 @@ void StringTestCase::WriteBuf()
         CPPUNIT_ASSERT_EQUAL(_T('a'), s[1]);
         CPPUNIT_ASSERT_EQUAL(_T('r'), s[2]);
         CPPUNIT_ASSERT_EQUAL(_T('r'), s[3]);
-        CPPUNIT_ASSERT_EQUAL(4u, s.length());
+        CPPUNIT_ASSERT_EQUAL((size_t)4, s.length());
 
         CPPUNIT_ASSERT_EQUAL( 0, wxStrcmp(_T("barr"), s) );
     }
